refactor: remove `Outgoing Mail` and `Incoming Mail`
