+++ mail/mail/doctype/dmarc_report/dmarc_report.json
@@ -6,7 +6,6 @@
  "engine": "InnoDB",
  "field_order": [
   "section_break_nb25",
-  "incoming_mail",
   "domain_name",
   "policy_published",
   "column_break_kyq2",
@@ -123,21 +122,12 @@
    "no_copy": 1,
    "read_only": 1,
    "search_index": 1
-  },
-  {
-   "fieldname": "incoming_mail",
-   "fieldtype": "Link",
-   "label": "Incoming Mail",
-   "no_copy": 1,
-   "options": "Incoming Mail",
-   "read_only": 1,
-   "search_index": 1
   }
  ],
  "in_create": 1,
  "index_web_pages_for_search": 1,
  "links": [],
- "modified": "2025-01-16 16:19:54.763220",
+ "modified": "2025-06-05 13:33:33.245733",
  "modified_by": "Administrator",
  "module": "Mail",
  "name": "DMARC Report",
@@ -157,6 +147,7 @@
    "write": 1
   }
  ],
+ "row_format": "Dynamic",
  "sort_field": "creation",
  "sort_order": "DESC",
  "states": [],
