--- mail/mail/doctype/incoming_mail/incoming_mail_list.js
@@ -1,33 +1,0 @@
-// Copyright (c) 2024, Frappe Technologies Pvt. Ltd. and contributors
-// For license information, please see license.txt
-
-frappe.listview_settings['Incoming Mail'] = {
-	refresh: (listview) => {
-		listview.page.add_inner_button('Fetch Emails', () => {
-			fetch_emails_from_clusters(listview)
-		})
-	},
-
-	get_indicator: (doc) => {
-		const status_colors = {
-			Draft: 'grey',
-			Submitted: 'blue',
-			Cancelled: 'red',
-		}
-		return [__(doc.status), status_colors[doc.status], 'status,=,' + doc.status]
-	},
-}
-
-function fetch_emails_from_clusters() {
-	frappe.call({
-		method: 'mail.tasks.enqueue_fetch_emails_from_clusters',
-		freeze: true,
-		freeze_message: __('Creating Job...'),
-		callback: () => {
-			frappe.show_alert({
-				message: __('{0} job has been created.', [__('Fetch Emails').bold()]),
-				indicator: 'green',
-			})
-		},
-	})
-}
