--- mail/mail/doctype/incoming_mail/test_incoming_mail.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2024, Frappe Technologies Pvt. Ltd. and Contributors
-# See license.txt
-
-# import frappe
-from frappe.tests.utils import FrappeTestCase
-
-
-class TestIncomingMail(FrappeTestCase):
-	pass
