+++ mail/mail/doctype/mail_cluster/mail_cluster.json
@@ -750,19 +750,9 @@
    "group": "Reference",
    "link_doctype": "Mail Backend Request",
    "link_fieldname": "backend_name"
-  },
-  {
-   "group": "Mail",
-   "link_doctype": "Incoming Mail",
-   "link_fieldname": "cluster"
-  },
-  {
-   "group": "Mail",
-   "link_doctype": "Outgoing Mail",
-   "link_fieldname": "cluster"
   }
  ],
- "modified": "2025-05-29 18:34:47.744984",
+ "modified": "2025-06-05 13:22:29.274007",
  "modified_by": "Administrator",
  "module": "Mail",
  "name": "Mail Cluster",
