--- mail/mail/doctype/mail_header/mail_header.json
@@ -1,44 +1,0 @@
-{
- "actions": [],
- "creation": "2024-02-13 10:55:02.513655",
- "doctype": "DocType",
- "editable_grid": 1,
- "engine": "InnoDB",
- "field_order": [
-  "key",
-  "column_break_79v8",
-  "value"
- ],
- "fields": [
-  {
-   "fieldname": "key",
-   "fieldtype": "Data",
-   "in_list_view": 1,
-   "label": "Key",
-   "reqd": 1
-  },
-  {
-   "fieldname": "value",
-   "fieldtype": "Text",
-   "ignore_xss_filter": 1,
-   "in_list_view": 1,
-   "label": "Value"
-  },
-  {
-   "fieldname": "column_break_79v8",
-   "fieldtype": "Column Break"
-  }
- ],
- "index_web_pages_for_search": 1,
- "istable": 1,
- "links": [],
- "modified": "2025-02-03 17:11:14.640642",
- "modified_by": "Administrator",
- "module": "Mail",
- "name": "Mail Header",
- "owner": "Administrator",
- "permissions": [],
- "sort_field": "modified",
- "sort_order": "DESC",
- "states": []
-}
