--- mail/mail/doctype/mail_header/mail_header.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2024, Frappe Technologies Pvt. Ltd. and contributors
-# For license information, please see license.txt
-
-# import frappe
-from frappe.model.document import Document
-
-
-class MailHeader(Document):
-	pass
