--- mail/mail/doctype/mail_recipient/mail_recipient.json
@@ -1,82 +1,0 @@
-{
- "actions": [],
- "creation": "2024-01-01 11:58:17.086749",
- "doctype": "DocType",
- "editable_grid": 1,
- "engine": "InnoDB",
- "field_order": [
-  "type",
-  "email",
-  "check_deliverability",
-  "display_name",
-  "column_break_iqoo",
-  "status",
-  "error_message"
- ],
- "fields": [
-  {
-   "fieldname": "column_break_iqoo",
-   "fieldtype": "Column Break"
-  },
-  {
-   "fieldname": "display_name",
-   "fieldtype": "Data",
-   "in_list_view": 1,
-   "label": "Display Name"
-  },
-  {
-   "fieldname": "status",
-   "fieldtype": "Select",
-   "in_list_view": 1,
-   "label": "Status",
-   "no_copy": 1,
-   "options": "\nBlocked\nSent",
-   "read_only": 1,
-   "search_index": 1
-  },
-  {
-   "default": "To",
-   "fieldname": "type",
-   "fieldtype": "Select",
-   "in_list_view": 1,
-   "label": "Type",
-   "options": "To\nCc\nBcc",
-   "reqd": 1,
-   "search_index": 1
-  },
-  {
-   "fieldname": "email",
-   "fieldtype": "Data",
-   "in_list_view": 1,
-   "label": "Email",
-   "options": "Email",
-   "reqd": 1,
-   "search_index": 1
-  },
-  {
-   "fieldname": "error_message",
-   "fieldtype": "Code",
-   "label": "Error Message",
-   "no_copy": 1,
-   "read_only": 1
-  },
-  {
-   "depends_on": "eval: doc.parenttype == \"Outgoing Mail\"",
-   "fieldname": "check_deliverability",
-   "fieldtype": "Button",
-   "label": "Check Deliverability"
-  }
- ],
- "index_web_pages_for_search": 1,
- "istable": 1,
- "links": [],
- "modified": "2025-02-19 19:18:44.222087",
- "modified_by": "Administrator",
- "module": "Mail",
- "name": "Mail Recipient",
- "owner": "Administrator",
- "permissions": [],
- "sort_field": "modified",
- "sort_order": "DESC",
- "states": []
-}
