+++ mail/mail/doctype/mail_settings/mail_settings.json
@@ -35,23 +35,7 @@
   "attachment_section",
   "max_attachments",
   "column_break_4lt9",
-  "max_attachment_size_mb",
   "max_total_attachments_size_mb",
-  "smtp_section",
-  "smtp_max_connections",
-  "smtp_cleanup_interval",
-  "column_break_9j6r",
-  "smtp_session_duration",
-  "smtp_inactivity_timeout",
-  "smtp_max_messages",
-  "limits_inbound_tab",
-  "imap_section",
-  "imap_max_connections",
-  "imap_cleanup_interval",
-  "column_break_adoa",
-  "imap_authenticated_timeout",
-  "imap_unauthenticated_timeout",
-  "imap_idle_timeout",
   "spamassassin_tab",
   "section_break_hgqa",
   "enable_spamd",
@@ -282,15 +266,6 @@
   },
   {
    "default": "10",
-   "description": "Maximum size allowed for each email attachment.",
-   "fieldname": "max_attachment_size_mb",
-   "fieldtype": "Int",
-   "label": "Maximum Attachment Size (MB)",
-   "non_negative": 1,
-   "reqd": 1
-  },
-  {
-   "default": "10",
    "description": "Maximum combined size for all attachments in an email.",
    "fieldname": "max_total_attachments_size_mb",
    "fieldtype": "Int",
@@ -316,46 +291,6 @@
    "reqd": 1
   },
   {
-   "description": "Changes may require a bench restart to take effect.",
-   "fieldname": "smtp_section",
-   "fieldtype": "Section Break",
-   "label": "SMTP"
-  },
-  {
-   "default": "300",
-   "description": "Maximum idle time before the session ends.",
-   "fieldname": "smtp_inactivity_timeout",
-   "fieldtype": "Int",
-   "label": "Inactivity Timeout (Seconds)",
-   "non_negative": 1,
-   "reqd": 1
-  },
-  {
-   "default": "600",
-   "description": "Maximum allowed time for an active session.",
-   "fieldname": "smtp_session_duration",
-   "fieldtype": "Int",
-   "label": "Session Duration (Seconds)",
-   "non_negative": 1,
-   "reqd": 1
-  },
-  {
-   "default": "10",
-   "description": "The maximum messages allowed per SMTP session.",
-   "fieldname": "smtp_max_messages",
-   "fieldtype": "Int",
-   "label": "Maximum Number of Messages",
-   "reqd": 1
-  },
-  {
-   "default": "3",
-   "description": "Max connections per host, port, and user. This limit applies per worker.",
-   "fieldname": "smtp_max_connections",
-   "fieldtype": "Int",
-   "label": "Maximum Number of Connections",
-   "reqd": 1
-  },
-  {
    "fieldname": "attachment_section",
    "fieldtype": "Section Break",
    "label": "Attachment"
@@ -366,10 +301,6 @@
    "label": "Outbound"
   },
   {
-   "fieldname": "column_break_9j6r",
-   "fieldtype": "Column Break"
-  },
-  {
    "fieldname": "limits_outbound_tab",
    "fieldtype": "Tab Break",
    "label": "Limits (Outbound)"
@@ -378,75 +309,6 @@
    "fieldname": "limit_outbound_message_section",
    "fieldtype": "Section Break",
    "label": "Message"
-  },
-  {
-   "default": "60",
-   "description": "The interval for cleaning up stale connections.",
-   "fieldname": "smtp_cleanup_interval",
-   "fieldtype": "Int",
-   "label": "Cleanup Interval (Seconds)",
-   "non_negative": 1,
-   "reqd": 1
-  },
-  {
-   "fieldname": "limits_inbound_tab",
-   "fieldtype": "Tab Break",
-   "label": "Limits (Inbound)"
-  },
-  {
-   "description": "Changes may require a bench restart to take effect.",
-   "fieldname": "imap_section",
-   "fieldtype": "Section Break",
-   "label": "IMAP"
-  },
-  {
-   "default": "3",
-   "description": "Max connections per host, port, and user. This limit applies per worker.",
-   "fieldname": "imap_max_connections",
-   "fieldtype": "Int",
-   "label": "Maximum Number of Connections",
-   "non_negative": 1,
-   "reqd": 1
-  },
-  {
-   "default": "60",
-   "description": "The interval for cleaning up stale connections.",
-   "fieldname": "imap_cleanup_interval",
-   "fieldtype": "Int",
-   "label": "Cleanup Interval (Seconds)",
-   "non_negative": 1,
-   "reqd": 1
-  },
-  {
-   "default": "1800",
-   "description": "Time an authenticated session can remain idle before termination.",
-   "fieldname": "imap_authenticated_timeout",
-   "fieldtype": "Int",
-   "label": "Authenticated Session Timeout (Seconds)",
-   "non_negative": 1,
-   "reqd": 1
-  },
-  {
-   "default": "60",
-   "description": "Time an unauthenticated session can remain inactive before termination.",
-   "fieldname": "imap_unauthenticated_timeout",
-   "fieldtype": "Int",
-   "label": "Unauthenticated Session Timeout (Seconds)",
-   "non_negative": 1,
-   "reqd": 1
-  },
-  {
-   "fieldname": "column_break_adoa",
-   "fieldtype": "Column Break"
-  },
-  {
-   "default": "1800",
-   "description": "Time a connection can stay idle in the IMAP IDLE state before termination.",
-   "fieldname": "imap_idle_timeout",
-   "fieldtype": "Int",
-   "label": "Idle Session Timeout (Seconds)",
-   "non_negative": 1,
-   "reqd": 1
   },
   {
    "default": "0",
@@ -557,7 +419,7 @@
  "index_web_pages_for_search": 1,
  "issingle": 1,
  "links": [],
- "modified": "2025-04-08 09:33:16.800557",
+ "modified": "2025-06-05 14:14:18.808516",
  "modified_by": "Administrator",
  "module": "Mail",
  "name": "Mail Settings",
