--- mail/mail/doctype/outgoing_mail/test_outgoing_mail.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2024, Frappe Technologies Pvt. Ltd. and Contributors
-# See license.txt
-
-# import frappe
-from frappe.tests.utils import FrappeTestCase
-
-
-class TestOutgoingMail(FrappeTestCase):
-	pass
