--- mail/mail/report/outbound_delay/outbound_delay.json
@@ -1,33 +1,0 @@
-{
- "add_total_row": 0,
- "columns": [],
- "creation": "2024-05-06 12:42:48.704926",
- "disabled": 0,
- "docstatus": 0,
- "doctype": "Report",
- "filters": [],
- "idx": 0,
- "is_standard": "Yes",
- "letterhead": null,
- "modified": "2025-02-04 21:01:00.357625",
- "modified_by": "Administrator",
- "module": "Mail",
- "name": "Outbound Delay",
- "owner": "Administrator",
- "prepared_report": 0,
- "ref_doctype": "Outgoing Mail",
- "report_name": "Outbound Delay",
- "report_type": "Script Report",
- "roles": [
-  {
-   "role": "System Manager"
-  },
-  {
-   "role": "Mail Admin"
-  },
-  {
-   "role": "Mail User"
-  }
- ],
- "timeout": 0
-}
