--- mail/mail/report/outgoing_mail_summary/outgoing_mail_summary.json
@@ -1,33 +1,0 @@
-{
- "add_total_row": 0,
- "columns": [],
- "creation": "2024-10-18 13:50:12.744195",
- "disabled": 0,
- "docstatus": 0,
- "doctype": "Report",
- "filters": [],
- "idx": 0,
- "is_standard": "Yes",
- "letterhead": null,
- "modified": "2025-02-04 21:05:41.074549",
- "modified_by": "Administrator",
- "module": "Mail",
- "name": "Outgoing Mail Summary",
- "owner": "Administrator",
- "prepared_report": 0,
- "ref_doctype": "Outgoing Mail",
- "report_name": "Outgoing Mail Summary",
- "report_type": "Script Report",
- "roles": [
-  {
-   "role": "System Manager"
-  },
-  {
-   "role": "Mail Admin"
-  },
-  {
-   "role": "Mail User"
-  }
- ],
- "timeout": 0
-}
